Put the timeseries tests in the same place
